drop the msp430.sh shim
Using the new asm-args field introduced in rust-lang/rust#38463, we can
fix the ABI problem in the target specification itself.

As the shim is gone, now the -mmcu flag is passed to the linker using
.cargo/config.

This commit also adds a build script. This build script will copy the
linker scripts from the ldscripts directory to $OUT_DIR. This (a) lets
you call the cargo command from a directory different that the root of
the Cargo project and (b) opens the possibility of using this crate as a
library.